### 2024-07-18 #207 @g4brielvs
BlackMarblePy Submission 
Submitting Author: Gabriel Stefanini Vicente (@g4brielvs)
All current maintainers: (@g4brielvs, @ramarty)
Package Name: BlackMarblePy
One-Line Description of Package: *Georeferenced Rasters and Statistics of Nightlights from NASA Black Marble*
Repository Link:  https://github.com/worldbank/blackmarblepy
Version submitted: [v2024.7.2](https://github.com/worldbank/blackmarblepy/releases/tag/v2024.07.2)
EiC: TBD
Version accepted: TBD

---

BlackMarblePy is a Python package that provides a simple way to use nighttime lights data from NASA’s Black Marble project. [Black Marble](https://blackmarble.gsfc.nasa.gov/) is a [NASA Earth Science Data Systems (ESDS)](https://www.earthdata.nasa.gov/) project that provides a product suite of daily, monthly and yearly global [nighttime lights](https://www.earthdata.nasa.gov/learn/backgrounders/nighttime-lights). This package automates the process of downloading all relevant tiles from the [NASA LAADS DAAC](https://www.earthdata.nasa.gov/eosdis/daacs/laads) to cover a region of interest, converting the raw files (in HDF5 format), to georeferenced rasters, and mosaicking rasters together when needed.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.
