### 2024-07-18 worldbank/blackmarblepy#86 @g4brielvs
pyOpenSci BlackMarblePy Submission
This issue is for tracking the https://github.com/pyOpenSci/software-submission/issues/207.

## Tasks
- [ ] https://github.com/pyOpenSci/software-submission/issues/207.
- [x] https://github.com/worldbank/blackmarblepy/pull/94
- [x] https://github.com/worldbank/blackmarblepy/pull/116
- [x] https://github.com/worldbank/blackmarblepy/pull/114
