### 2024-07-22 #207 @cmarmo
Hi @g4brielvs , thank you for submitting BlackMarblePy to pyOpenSci.
I'm Chiara, currently serving as Editor In Chief. I  just want to let you know that I'm going to perform the first checks on the package ASAP, hopefully by the end of the month.
Thanks for your patience!
