### 2024-07-22 #207 @g4brielvs
Hi Chira (@cmarmo)! Thank you for your message and for taking the time to review [BlackMarblePy](https://worldbank.github.io/blackmarblepy). I appreciate the update and look forward to any feedback you may have.
