### 2024-07-31 #207 @cmarmo
Hi @g4brielvs ! Thank you for submitting your package for pyOpenSci review.
Sorry for taking so much time to come back to you.

Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

---
---

Below my comments about the unchecked boxes up there, and some others related to more general questions.

- While the examples in the documentation are valuable answers to some of the questions in the submission form, the audience of the package and the possible scientific applications haven't been detailed. Also a comparison with other similar packages (if any) is missing. Do you mind adding some details about that in the submission form or as a follow-up discussion in the issue?
- The documentation contains an API section but I couldn't easily find it because its title is "blackmarble package" rather than eg "blackmarble API" or "blackmarble package API", do you mind making the section title more explicit?
- The Code of conduct has been added via a github template? The file itself is missing, then the link in the [CONTRIBUTING.md](https://github.com/worldbank/blackmarblepy/blob/main/CONTRIBUTING.md) gives a "404 not found".
- While looking for the API chapter in the documentation sources, I realized that you are using google analytics to monitor the documentation. Please note that google analytics needs cookie consent for European General Data Protection Regulation and specific configuration for California Consumer Privacy Act. Before starting the review do you mind adding a cookie consent banner or switching to another service GDPR and CCPA compliant (eg plausible https://plausible.io/)?
-  I couldn't find your pre-review survey... perhaps something went wrong during its submission. Do you mind refilling it? This helps us track submission and improve our peer review process.
