### 2024-07-31 0246547
Fix CONTRIBUTING.md links
