### 2024-07-31 08b7256
Fixes for pyOpenSci Submission (#94)
* Remove GA

* Fix CONTRIBUTING.md links

* Fix CONTRIBUTING.md links

* Move CONTRIBUTING.md

* Update Documentation title
