+++ docs/_config.yml
@@ -2,7 +2,7 @@
 title:
 author: Development Data Group
 logo: images/logo.png
-only_build_toc_files: true
+only_build_toc_files: false
 
 repository:
   url: https://github.com/worldbank/blackmarblepy
