### 2024-07-31 #207 @g4brielvs
@cmarmo Thank you **so** much for reviewing BlackMarblePy. The issues raised on your comments have been [addressed](https://github.com/worldbank/blackmarblepy/pull/94). Please let us know if you have any additional feedback or there is anything I can do to help.
