### 2024-08-01 #207 @g4brielvs
> Thank you @g4brielvs ! I believe we are ready to seek for an editor. I wonder if you can tag this last version and update the "submitted version" in the description, so your last modifications are transparent for the reviewers. Thank you so much!

@cmarmo Thanks **so** much! And absolutely. I just create the [v2024.08.1](https://github.com/worldbank/blackmarblepy/releases/tag/v2024.08.1) release.
