### 2024-08-15 #207 @cmarmo
Hello @g4brielvs, I am glad to announce that @yeelauren kindly accepted to serve as editor for the BlackMarblePy submission. Thank you so much Lauren!

I'm letting Lauren introduce herself here and wishing to you all a happy review process! :tada:
