+++ .pre-commit-config.yaml
@@ -34,7 +34,7 @@ repos:
         name: codespell
         description: Checks for common misspellings in text files
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.5.6
+    rev: v0.6.3
     hooks:
       - id: ruff
         types_or: [python, pyi, jupyter]
