### 2024-09-03 #207 @g4brielvs
Thank you for the update @cmarmo @lwasser! Welcome, @yeelauren, and thank you for taking on the role of editor for the [BlackMarblePy submission](https://github.com/pyOpenSci/software-submission/issues/207). I’m looking forward to working with you throughout the review process. If there’s anything you need from me or any way I can assist, please don’t hesitate to reach out.
