### 2024-09-05 worldbank/blackmarblepy#96 @Skerre
@ashbate Dear ashbate,

thank you for the effort of resolving this. However, I tried your solution and it did not work yet. I added the timeout to the beginning of the script (also tried other places, within the functions) and literally replaced my _download function with yours. It does not really seem to make a timeout and goes into the same error as before.

![image](https://github.com/user-attachments/assets/ceb7ebe7-abdd-4b35-9989-f96fd74fd4dd)

![image](https://github.com/user-attachments/assets/2ee5dc36-04f2-494d-a6de-45d58085d15f)
