### 2024-09-05 worldbank/blackmarblepy#96 @Skerre
@ashbate I can share more details on what I am doing if you want
