### 2024-10-16 #207 @yeelauren
## Editor response to review:

---

## Editor comments
:wave: Hi @petegadomski and @ehinman! Thank you for volunteering to review
for pyOpenSci!  I'm excited to have you both join us :partying_face:

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: <Insert deadline DATE HERE>
---

Reviewers:@petegadomski and @ehinman
Due date: Nov 6th, 2024
