### 2024-11-05 #207 @ehinman
I am having the same issue as mentioned here: https://github.com/worldbank/blackmarblepy/issues/103. I'm not sure how to get around it. I tried using my own gdf shape, as well as the example given in [Using `BlackMarblepy`](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html), and both are returning the same error. Any suggestions? I tried the solutions reported in issue 103, but they don't seem to be working for me.
