### 2024-11-06 #207 @gadomski
- [ ] **Vignette(s)** demonstrating major functionality that runs successfully locally.[^1]
- ~[ ] **Examples** for all user-facing functions.~ (see notes for editors)

    - [x] Continuous integration ~and test coverage,~ (see notes for editors)

- [ ] Any additional setup required to use the package (authentication tokens, etc.)[^3]
    - [ ] Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear)[^1]

- [ ] **Functionality:** Any functional claims of the software been confirmed.[^4]
- [ ] **Performance:** Any performance claims of the software been confirmed.[^4]
- [ ] **Automated tests:**[^2][^5]
- [ ] **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)[^6]

| Date | Hours |
| -- | -- |
| 2024-11-05 | 1 |

---

#### Review notes

I've included my comments to the authors as footnotes. For the editors:

-  🤔 I'm not sure I agree with the requirement to have examples for all user-facing functions. It seems overly onerous and fragile (unless you're working in an environment that executes all examples in docstrings, which isn't all Python projects).
- I don't agree w/ the coverage badge requirement, but that's more of a personal thing — I find coverage a bit of a false flag at time that's worth more hassle than you get.

[^1]: The README vignette is non-trivial to execute as a new user. How do I create the geodata frame? Why I can't I use a **shapely** shape? What does a `bearer` token look like?
[^2]: The README doesn't include instructions on how to run tests. Should I run **pytest**? Are there other things I need to install to run tests?
[^3]: The README should call out what token you're expected to get+use (earthdata)
[^4]: I also ran into https://github.com/worldbank/blackmarblepy/issues/96 when doing a local version of https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html. I can't check off "functionality" if it doesn't work that easily, sorry 😞. I'll also skip performance until functionality is resolved.
[^5]: The tests don't actually exercise the package.
[^6]: CI doesn't actually block on **ruff** failures.
