### 2024-11-06 b7e6c72
fix: some doc fixups
These arose while doing my pyOpenSci review:
https://github.com/pyOpenSci/software-submission/issues/207#issuecomment-2458503457
