+++ docs/SECURITY.md
@@ -1,17 +1,1 @@
-# Security Policy
-
-## Supported Versions
-
-| Version   | Supported          |
-| ----------| ------------------ |
-| > 2024.07.x | :white_check_mark: |
-| < 1.0.x   | :x:                |
-
-## Reporting a Vulnerability
-
-If you discover a vulnerability in [BlackMarblePy](https://github.com/worldbank/blackmarblepy), we encourage you to [report it](https://github.com/worldbank/blackmarblepy/issues/new/choose) to us as soon as possible. Please include as much detail as possible, including:
-
-- A description of the vulnerability
-- Steps to reproduce the issue
-- Potential impact
-- Any available proof-of-concept or exploit code
+../SECURITY.md
