### 2024-11-06 #207 @ehinman
- [x] Any additional setup required to use the package (authentication tokens, etc.)*

---

- I do not have access to the following link in the README: https://www.earthdata.nasa.gov/topics/human-dimensions/nighttime-lights
- Might be helpful to link to products (`product_id`) somewhere, since it wasn't immediately clear to me what the alphanumeric codes meant in the function calls: https://blackmarble.gsfc.nasa.gov/#product
- *I figured out how to get a NASA API bearer token and apply it as an environment variable using `os.environ['BLACKMARBLE_TOKEN'] = ######` from [Stack Overflow](https://stackoverflow.com/questions/5971312/how-to-set-environment-variables-in-python), but the [Debian](https://wiki.debian.org/EnvironmentVariables) link was not straightforward. More guidance would be helpful.
- **BIGGEST ISSUE:** I was unable to run any of the functions due to an error highlighted in [this issue](https://github.com/worldbank/blackmarblepy/issues/96). This greatly impacted my ability to review the package.
- Can you provide an example for each function in the documentation? Or at least in a vignette? I understand this is hard with the API token requirement, but just seeing how a function is parameterized in an example would be helpful.
- Is `BlackMarbleDownloader` intended to be used separate from `bm_raster` and `bm_extract`? It might be helpful to provide more details on its intended use in the documentation. For example: "A function utilized by `bm_raster` and `bm_extract`."
- Can you provide documentation for `blackmarble.raster.transform`?
- Not sure where to view test coverage %.
- I don't see which python versions are supported. I tried to run it with `python 3.12.0`
- Might be nice to link to the R version of this package as well.
- The test coverage looks extremely limited. I ran `pytest` in the tests folder on the 3 existing tests...are there more that I'm missing?
- There are several lines greater than 79 characters, but this doesn't concern me much.
