### 2024-11-27 worldbank/blackmarblepy#96 @Abdulazizbek
I'm facing same issue. I have followed @ashbate and @koichisato-dev 's modifications, but still issue is not solved in my case.

However, I found another possible cause of problem in their starter [guideline](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html) page (2nd step):

> Please be aware that the “Affiliation” information on your [Earthdata profile](https://urs.earthdata.nasa.gov/profile/edit) is mandatory. Without this information, the [NASA Earthdata token](https://urs.earthdata.nasa.gov/documentation) will be invalid, which may result in an error (i.e., OSError: Unable to synchronously open file (file signature not found))

![image](https://github.com/user-attachments/assets/a05a477c-cafa-4cb7-b6ab-1c4b4ffb4533)

I have checked profile information and bearer status (I don't know how to check **is it working or not with this bearer token**) according to above guideline, but still problem is not solved. **If anyone has dealt with this problem, a guide to overcoming it would be very helpful.**

Thank you beforehand.
