### 2024-11-28 worldbank/blackmarblepy#96 @koichisato-dev
@Abdulazizbek I cannot see your current code, so I am not sure why the error is occurring, but I suspect you might not have set up your .env file correctly or loaded your token properly. Could you try the following code?

<setting in key.env>
BLACK_MARBLE_BEAPER_TOKEN = your token

```
load_dotenv('your_path/key.env')
bearer = os.getenv('BLACK_MARBLE_BEARER_TOKEN')

if not bearer:
    raise ValueError('No bearer token found in environment variables')

 ntl_monthly_mean = bm_extract(gdf,
                              product_id="VNP46A3",
                              date_range=pd.date_range('2014-01-01', '2024-08-01', freq='MS'),
                              output_directory='your directory',
                              output_skip_if_exists=False,
                              variable='AllAngle_Composite_Snow_Free',
                              bearer=bearer)
```
