### 2024-12-03 worldbank/blackmarblepy#96 @Abdulazizbek
@koichisato-dev Thank you for your explanation. I have checked with your provided code and my token is working properly (expiry  date is normal). But, still getting some issues like below:
![Screenshot from 2024-12-03](https://github.com/user-attachments/assets/2a0a832c-fe24-4e92-b1fd-bc1c97b65003)
![Screenshot from 2024-12-03_1](https://github.com/user-attachments/assets/3c60cb14-8d47-4a41-859e-fddd93b36abc)
![Screenshot from 2024-12-03_2](https://github.com/user-attachments/assets/9d45e4b7-b5f3-4e65-9bfc-cfb64ded0c7d)

I'm not sure where is going wrong. If someone handled successfully, source code would be helpful.
Thank you beforehand.
