### 2024-12-04 worldbank/blackmarblepy#96 @koichisato-dev
@Abdulazizbek If that's the case, the 'output_directory' path in bm_extract is incorrect. Please ensure you specify the correct path.
