### 2024-12-04 worldbank/blackmarblepy#96 @Abdulazizbek
@koichisato-dev I have double checked parameters and I have checked images saved in 'output_directory'
