### 2024-12-04 worldbank/blackmarblepy#96 @koichisato-dev
@Abdulazizbek okay, great! Then, continue the tutorial
