### 2025-01-09 worldbank/blackmarblepy#96 @msoltadeo
I am having the same issues. The h5 are being downloaded with 0 bytes. However, when I follow the url for downloading the data, I can download it with no problem.
