### 2025-02-13 #207 @g4brielvs
> @g4brielvs happy new year ! Just checking in on this review and progress :)

@yeelauren Thank you for checking in and happy new year. A new release should be coming soon, which should address the issues raised. Would you be able to accommodate this timeframe?
