### 2025-06-02 #207 @g4brielvs
@ehinman Thank you **so** much for reviewing [blackmarblepy](https://worldbank.github.io/blackmarblepy)! Apologies for the interregnum. A new release [v2025.06.1](https://github.com/worldbank/blackmarblepy/releases/tag/v2025.06.1) is now available.
