### 2025-06-06 #207 @g4brielvs
@ehinman @gadomski @yeelauren Thank you so much for the thorough and thoughtful feedback. It is incredibly helpful as we continue to improve [blackmarblepy](https://worldbank.github.io/blackmarblepy)! A new release [v2025.06.2](https://pypi.org/project/blackmarblepy/) is now available.

### On @gadomski (thanks so much!):

> I'm not sure I agree with the requirement to have examples for all user-facing functions. It seems overly onerous and fragile (unless you're working in an environment that executes all examples in docstrings, which isn't all Python projects).

I am with you on this. I agree that requiring examples for every function can be overly burdensome. Unfortunately, it was a requirement at the time I was working, so we aimed to meet that standard to align with internal review and reproducibility expectations. Definitely open to a more balanced approach moving forward.

### On @ehinman's review (thanks so much!):
>     * I do not have access to the following link in the README: https://www.earthdata.nasa.gov/topics/human-dimensions/nighttime-lights

It appears to be working now. Could you confirm if you are able to access it on your end as well?

>     * Might be helpful to link to products (`product_id`) somewhere, since it wasn't immediately clear to me what the alphanumeric codes meant in the function calls: https://blackmarble.gsfc.nasa.gov/#product

We have added the [documentation](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.types.Product) and to the [docstrings](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.raster.bm_raster).

>     * *I figured out how to get a NASA API bearer token and apply it as an environment variable using `os.environ['BLACKMARBLE_TOKEN'] = ######` from [Stack Overflow](https://stackoverflow.com/questions/5971312/how-to-set-environment-variables-in-python), but the [Debian](https://wiki.debian.org/EnvironmentVariables) link was not straightforward. More guidance would be helpful.

You are absolutely right that the Debian environment variable documentation can be a bit opaque. Our intention was to suggest setting the token outside the Python script as a best practice for managing secrets, but to clarify, the function itself simply expects a string, so you are free to pass the token directly if needed.

We will update the [documentation](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html#generate-nasa-earthdata-token) to make this distinction clearer.

>     * **BIGGEST ISSUE:** I was unable to run any of the functions due to an error highlighted in [this issue](https://github.com/worldbank/blackmarblepy/issues/96). This greatly impacted my ability to review the package.

Thank you for flagging this. The issue has been resolved as of [v2025.06.1](https://pypi.org/project/blackmarblepy/2025.6.1/). Thank you for your patience. If you are still encountering errors, please let us know.

>     * Can you provide an example for each function in the documentation? Or at least in a vignette? I understand this is hard with the API token requirement, but just seeing how a function is parameterized in an example would be helpful.

We have provided examples as Jupyter notebook, which you can view or run directly, including in Google Colab, on [this page](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html#examples).

>     * Is `BlackMarbleDownloader` intended to be used separate from `bm_raster` and `bm_extract`? It might be helpful to provide more details on its intended use in the documentation. For example: "A function utilized by `bm_raster` and `bm_extract`."

Thanks for catching this. You are correct. `BlackMarbleDownloader` is intended as a lower-level utility used by `bm_raster` and `bm_extract`. We will update the documentation to explicitly state this.

>     * I don't see which python versions are supported. I tried to run it with `python 3.12.0`

The supported Python versions are listed on [PyPI](https://pypi.org/project/blackmarblepy/) and specified in the `pyproject.toml`. To make this clearer, we have now added a Python version badge to the documentation and README so users can quickly see compatibility.

>     * Might be nice to link to the R version of this package as well.

Excellent suggestion. We already list the [blog post](https://blogs.worldbank.org/en/opendata/illuminating-insights-harnessing-nasas-black-marble-r-and-python-packages). We added a link to the [R version of the package](https://github.com/worldbank/blackmarbler) in both the [README and the documentation](https://worldbank.github.io/blackmarblepy/README.html#related-projects).

>     * Can you provide documentation for `blackmarble.raster.transform`?
>     * Not sure where to view test coverage %.
>     * The test coverage looks extremely limited. I ran `pytest` in the tests folder on the 3 existing tests...are there more that I'm missing?

You are right that current test coverage is minimal. We are working on expanding this, but it is a bit tricky due to the dependency on NASA. We are exploring ways to mock to improve testability.

>     * There are several lines greater than 79 characters, but this doesn't concern me much.

We appreciate the note. We aim to follow PEP8 where feasible, though we sometimes allow longer lines when it improves clarity.

@ehinman @gadomski **Thank you so much** for your contributions and patience! Please let us know if there is anything else we can do to support the review.
