### 2025-10-07 #207 @lwasser
Hi Friends. It looks like this package review has stalled. What do we need to do to move things forward? @yeelauren is going to be taking on the EiC role so I wonder if we want to have a second editor help get it across the finish line? Lauren and everyone here, what do you prefer here, and how can I help move this package forward?
