### 2025-10-08 #207 @gadomski
Thanks for the ping! I had this in my backlog but it got dropped over the summer. I can probably get to a re-review sometime in October ... hopefully that's not too slow?
