### 2025-10-13 #207 @yeelauren
Hi @ehinman and @gadomski,

Thank you both for your reviews so far, and a special thanks to @g4brielvs for releasing a new version that addresses your comments!

Does the end of October 👻 🕸️ sound like a reasonable time frame for you to evaluate whether the changes effectively address the issues raised in your reviews?
