### 2025-11-02 #207 @g4brielvs
Hi @lwasser @yeelauren @ehinman @gadomski, 🎃 Happy (belated) Halloween) !

Just wanted to share a quick update. [BlackMarblePy v2025.11.1](https://pypi.org/project/blackmarblepy/2025.11.1) adds support to Collection 2 ("5200"), which should resolve any potential errors for date ranges starting in 2025.

We really appreciate your feedback and the time you have taken to review this package. It is used extensively by the World Bank and has been recognized by the Black Marble team (as listed on their [tools page](https://blackmarble.gsfc.nasa.gov/Tools.html)).

We're excited to continue expanding its reach to the broader community.
