### 2025-11-04 #207 @gadomski
Just an update here, I wasn't able to get to this in October, but I'm still hoping to get to re-review in the next week or so. Apologies for the delay 🙇🏼.
