### 2025-11-10 #207 @gadomski
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [ ] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally. (see note at bottom of this comment)
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

- ~[ ] **Performance:** Any performance claims of the software been confirmed.~ n/a, as far as I could tell there weren't any significant performance claims

---

## Notes

- I wasn't able to get the quickstart working: https://github.com/worldbank/blackmarblepy/issues/137
- There's some funny image rendering on the documentation page: <img alt="Image" width="400" src="https://github.com/user-attachments/assets/443e65d9-fee0-42cc-a6db-95687b7bb594" />
- Tests are a little light, but since it's a client to an external service I think that's ok
