### 2025-11-13 #207 @g4brielvs
@gadomski Thank you so much for your comments and reviewing! We published [v2025.11.2](https://pypi.org/project/blackmarblepy/2025.11.2/).

We have an [open issue](https://github.com/worldbank/blackmarblepy/issues/21) for improving testing. As you mentioned, it is tricky to test comprehensively considering that it is a client, but we will look into mocking and expand testing.
