### 2025-11-25 #207 @elisehinman
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [ ] **Examples** for all user-facing functions. *Unclear to me if this has been achieved*
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem. *Not applicable*
- [x] Citation information *But see comment below on some hanging markdown/docstring text*

The package structure should follow the general community best practices. In general, please consider whether:

- [ ] Package documentation is clear and easy to find and use. *I must rely on the vignettes because I do not think I could figure out how to use the package with the API reference alone. Would like more simple examples.*

- [x] **Installation:** Installation succeeds as documented. *Once I had geopandas installed*

---

- why is uv recommended? It might be nice to have a statement of reason.
- When I ran `pip install blackmarblepy`, I received the following error:
```

  × Getting requirements to build wheel did not run successfully.
      WARNING:root:Failed to get options via gdal-config: [Errno 2] No such file or directory: 'gdal-config'
      ERROR: A GDAL API version must be specified. Provide a path to gdal-config using a GDAL_CONFIG environment variable or use a GDAL_VERSION environment variable.

ERROR: Failed to build 'rasterio' when getting requirements to build wheel"
```
I had to install `geopandas` first before `blackmarblepy`.
- Can you link to an actual demonstration of how to add a variable to your environment? I didn't find the link super helpful.
- Are there any "flavors" of contribution that you'd prefer/like to avoid? For example, I'd assume you want contributions that improve/build off the core set of functions already in the package, as opposed to someone creating customizations that may be narrow in utility, or adding a new non-Black Marble service.
- I'm seeing some weird hanging docstrings/markdown syntax at the end of the Citation:
"{cite:empty}blackmarblepy

:filter: docname in docnames
:style: plain"
- Broken link to "official documentation" under "Full API Reference" heading.
- Would love to see descriptions (and links) of the three vignettes from the documentation website available on the README.md.
- I am confused by the API reference to the core module. It might be nice to see how modules and functions interact with each other. For example, I do not understand how the `extract` function in the core module is different from the `extract` module. And why doesn't `close` have its own module?  Or `collate_tiles`? It would be nice to see examples of how all these functions are used in practice.
- It might be nice to have a `requirements.txt` or `environment.yml` that has all the packages needed to run the vignettes. I'm running into package conflicts, specifically with `contextily`.
- I am trying to follow the "Using BlackMarblePy" vignette. I can enter my token successfully, but when I try to run the following code, I get an error:
```
VNP46A2 = bm.raster(
    gdf,
    product_id="VNP46A2",
    date_range="2024-03-21",
)
VNP46A2
RuntimeError                              Traceback (most recent call last)
----> 1 VNP46A2 = bm.raster(
      2     gdf,
      3     product_id="VNP46A2",
      4     date_range="2024-03-21",
      5 )
      6 VNP46A2

File /opt/homebrew/Caskroom/miniforge/base/envs/blackmarble1/lib/python3.14/site-packages/pydantic/_internal/_validate_call.py:39, in update_wrapper_attributes.<locals>.wrapper_function(*args, **kwargs)
     37 @functools.wraps(wrapped)
     38 def wrapper_function(*args, **kwargs):
---> 39     return wrapper(*args, **kwargs)

File /opt/homebrew/Caskroom/miniforge/base/envs/blackmarble1/lib/python3.14/site-packages/pydantic/_internal/_validate_call.py:136, in ValidateCallWrapper.__call__(self, *args, **kwargs)
    133 if not self.__pydantic_complete__:
    134     self._create_validators()
--> 136 res = self.__pydantic_validator__.validate_python(pydantic_core.ArgsKwargs(args, kwargs))
    137 if self.__return_pydantic_validator__:
    138     return self.__return_pydantic_validator__(res)

File /opt/homebrew/Caskroom/miniforge/base/envs/blackmarble1/lib/python3.14/site-packages/blackmarble/core.py:553, in BlackMarble.raster(self, gdf, product_id, date_range, variable)
    548 # Download NASA Black Marble tiles for the specified region and dates and
    549 # returns a mapping of file paths grouped by date for further processing.
...
    470     )
    473     self._active = False

RuntimeError: Attempted to exit a cancel scope that isn't the current tasks's current cancel scope
```
Let me know if I should make this an issue. I can share my environment specifications if that would help.

I will wait to continue my review until I can get this resolved.
