### 2025-11-26 #207 @g4brielvs
@elisehinman Thank you **so** much for your review of [BlackMarblePy](https://github.com/worldbank/blackmarblepy)! Thanks to your feedback, I created a [few issues to work on](https://github.com/worldbank/blackmarblepy/issues) and I added comments to the items you raised below:

>     * why is uv recommended? It might be nice to have a statement of reason.

That's a good point. I think it is best to defer to the [uv project linked](https://docs.astral.sh/uv/#highlights), which provides a more complete overview. What do you think?

>     * When I ran `pip install blackmarblepy`, I received the following error:
> ```
> error: subprocess-exited-with-error
>
>   × Getting requirements to build wheel did not run successfully.
>   │ exit code: 1
>   ╰─> [2 lines of output]
>       WARNING:root:Failed to get options via gdal-config: [Errno 2] No such file or directory: 'gdal-config'
>       ERROR: A GDAL API version must be specified. Provide a path to gdal-config using a GDAL_CONFIG environment variable or use a GDAL_VERSION environment variable.
>       [end of output]
>
>   note: This error originates from a subprocess, and is likely not a problem with pip.
> ERROR: Failed to build 'rasterio' when getting requirements to build wheel"
> ```
>
> I had to install `geopandas` first before `blackmarblepy`.

At this time, [BlackMarblePy](https://pypi.org/project/blackmarblepy/)'s dependencies are [not yet compatible](https://github.com/geopandas/geopandas/pull/3659) with [python3.14](https://www.python.org/downloads/release/python-3140/), which seems to be causing the error above on your environment. [GeoPandas](https://geopandas.org/en/stable/getting_started/introduction.html) is [listed a dependency](https://github.com/worldbank/blackmarblepy/blob/1bdad1af8e218feb7402146bdc2325ac3a224bfd/pyproject.toml#L34), so you wouldn't have to install it on a compatible environment.

While the [supported versions are documented](https://pypi.org/project/blackmarblepy), we considered adding a `<3.14 `version constraint explicitly, until that version is compatible. I created an [issue to keep track of this](https://github.com/worldbank/blackmarblepy/issues/143).

>     * Can you link to an actual demonstration of how to add a variable to your environment? I didn't find the link super helpful.

Thanks! There are actually several ways to add an environment variable and the best approach can vary depending on the user's preferences. It's often simplest to refer users to standard documentation or examples.

That said, there's also the [option of passing the token as argument](https://github.com/worldbank/blackmarblepy?tab=readme-ov-file#usage), which can be helpful if the user isn't familiar with environment variables.

>     * Are there any "flavors" of contribution that you'd prefer/like to avoid? For example, I'd assume you want contributions that improve/build off the core set of functions already in the package, as opposed to someone creating customizations that may be narrow in utility, or adding a new non-Black Marble service.

Honestly, all contributions are welcome. We appreciate any effort to improve the project. During the normal review process, PRs will go through triage, which helps us determine whether something fits well within the scope of the package or should be discussed further. What do you think?

>     * I'm seeing some weird hanging docstrings/markdown syntax at the end of the Citation:
>       "{cite:empty}blackmarblepy
>
>
> :filter: docname in docnames :style: plain"

These are for [Jupyter Book](https://jupyterbook.org/stable/]), which is used to build the [documentation page](https://worldbank.github.io/blackmarblepy/README.html)

>     * Broken link to "official documentation" under "Full API Reference" heading.

The "API Reference" section is built automatically based on the package's docstrings on the [documentation page](https://worldbank.github.io/blackmarblepy/README.html), where the link isn't broken:

- https://worldbank.github.io/blackmarblepy/api/blackmarble.html

However, itcould be good to replace with the link, so it isn't broken on GitHub. Thanks!

>     * Would love to see descriptions (and links) of the three vignettes from the documentation website available on the README.md.

That's a good point! There's a link to ["more examples"](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html) below the [vignette on the README](https://worldbank.github.io/blackmarblepy/README.html#usage), but we can make it better.

>     * I am confused by the API reference to the core module. It might be nice to see how modules and functions interact with each other. For example, I do not understand how the `extract` function in the core module is different from the `extract` module. And why doesn't `close` have its own module?  Or `collate_tiles`? It would be nice to see examples of how all these functions are used in practice.

The submodules [`raster`](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html) and [`extract`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#module-blackmarble.raster) are present currently to ensure backward compatibility. Initially, we started the library adopting a functional-oriented approach (i.e., [`bm_raster`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.raster.bm_raster)), then moved to a object-oriented approach (i.e., [`BlackMarble`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.core.BlackMarble)). They accomplish the same. In matter of fact, [`bm_raster`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.raster.bm_raster) calls  [`BlackMarble`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.core.BlackMarble). In the future, we might consider deprecation.

The functions [`close`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.core.BlackMarble.close) and [`collate_tiles`](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#blackmarble.core.BlackMarble.collate_tiles) are instance methods, their respective sections are available on the [package documentation](https://worldbank.github.io/blackmarblepy/api/blackmarble.html#).

>     * It might be nice to have a `requirements.txt` or `environment.yml` that has all the packages needed to run the vignettes. I'm running into package conflicts, specifically with `contextily`.

Admittedly a bit implicit, optional dependencies used for the [example notebook](https://github.com/worldbank/blackmarblepy/blob/main/notebooks/blackmarblepy.ipynb) can be installed with:

```bash
uv add "blackmarblepy[examples]"
```

These include dependencies not required for core functionality, like [Folium](https://python-visualization.github.io/folium/latest/) for [maps](https://worldbank.github.io/blackmarblepy/notebooks/blackmarblepy.html#define-region-of-interest).

>     * I am trying to follow the "Using BlackMarblePy" vignette. I can enter my token successfully, but when I try to run the following code, I get an error:
>
>
> ```
> VNP46A2 = bm.raster(
>     gdf,
>     product_id="VNP46A2",
>     date_range="2024-03-21",
> )
> VNP46A2
> ---------------------------------------------------------------------------
> RuntimeError                              Traceback (most recent call last)
> Cell In[5], line 1
> ----> 1 VNP46A2 = bm.raster(
>       2     gdf,
>       3     product_id="VNP46A2",
>       4     date_range="2024-03-21",
>       5 )
>       6 VNP46A2
>
> File /opt/homebrew/Caskroom/miniforge/base/envs/blackmarble1/lib/python3.14/site-packages/pydantic/_internal/_validate_call.py:39, in update_wrapper_attributes.<locals>.wrapper_function(*args, **kwargs)
>      37 @functools.wraps(wrapped)
>      38 def wrapper_function(*args, **kwargs):
> ---> 39     return wrapper(*args, **kwargs)
>
> File /opt/homebrew/Caskroom/miniforge/base/envs/blackmarble1/lib/python3.14/site-packages/pydantic/_internal/_validate_call.py:136, in ValidateCallWrapper.__call__(self, *args, **kwargs)
>     133 if not self.__pydantic_complete__:
>     134     self._create_validators()
> --> 136 res = self.__pydantic_validator__.validate_python(pydantic_core.ArgsKwargs(args, kwargs))
>     137 if self.__return_pydantic_validator__:
>     138     return self.__return_pydantic_validator__(res)
>
> File /opt/homebrew/Caskroom/miniforge/base/envs/blackmarble1/lib/python3.14/site-packages/blackmarble/core.py:553, in BlackMarble.raster(self, gdf, product_id, date_range, variable)
>     548 # Download NASA Black Marble tiles for the specified region and dates and
>     549 # returns a mapping of file paths grouped by date for further processing.
> ...
>     470     )
>     472 try:
>     473     self._active = False
>
> RuntimeError: Attempted to exit a cancel scope that isn't the current tasks's current cancel scope
> ```

Thank you for testing. Downgrading from [python3.14](https://www.python.org/downloads/release/python-3140/) should address the issue.
