### 2025-11-27 #207 @elisehinman
Hm. I keep running into the same error when I try to `pip install blackmarblepy`. I have tried creating new environments with python 3.13.9 and 3.12.12. Both times the same issue. However, when I install `geopandas` first, I can successfully download data using `blackmarblepy` on python 3.12.12!
