### 2025-11-27 #207 @elisehinman
Additional comments:

- On GitHub, the broken full API Reference link goes to: https://github.com/worldbank/blackmarblepy/blob/main/api/blackmarble.rst, which doesn't exist and throws a 404. Agree that it would be nice to have it link to the actual html.
- The vignette itself doesn't show the packages imported to run the vignette. I went to one of the online notebooks to see the imported packages.
- While I appreciate that it's easy with uv to install all the packages needed to run the examples, you could hold non-uv users hands (like mine, as someone unfamiliar with uv and these datasets) a little more in the notebook and explicitly give them some options for building their environment most efficiently.
- Can you give a few more details on the usage of `date_range` in `bm.raster()`? In the "Using BlackMarblePy" vignette, it's interesting that you use the same date to grab daily, monthly, and annual rasters. Does `blackmarblepy` behind the scenes extract the month or year from the `date_range`? You could explain this either in the vignette itself or in the API reference.
- On the "Using BlackMarblePy" vignette, I didn't realize the "daily raster stack" would be 11.4 GB. Might be nice to warn users the size in case they're working with limited space. Or, is it possible to more closely specify what parameters come back in the xarray?
- Can you add more comments in the code cell blocks of the vignette describing what you're doing? For example, here: `# Plot NTL radiance
r_monthly["NearNadir_Composite_Snow_Free"].sum(dim=["x", "y"]).plot(ax=ax)` Why does summing them give nighttime light radiance?
- The monthly nighttime light radiance plot I think should say "average", not "aggregate" in the title. This confused me, initially. I'm assuming the "yearly" plots are showing an average, too, right? Also, can you show the yearly radiance plots with just the year in the title of each plot, without the "time = YYYY-01-01"? That is a little confusing, since it's a yearly average, right?
- Related to https://github.com/worldbank/blackmarblepy/issues/142, can you make the error messages more informative when downloads fail due to permissions, expired token, or something else? I think the error message before I had proper permissions suggested that maybe my Token was expired, though I had just made a new one this session.

This is SO exciting to be able to actually use and view these datasets! Thank you for the opportunity to re-review. I have completed my re-review with vast improvements in my ability to use the package.
