### 2026-04-13 #207 @g4brielvs
Hi! Thank you for checking up and apologies for the delay. We are reviewing the package to upgrade to the version 2 that NASA released. If possible, we would appreciate your review and feedback in 2-3 weeks when all changes are merged. We appreciate your input and patience!
