### 2026-04-13 #207 @eliotwrobson
@g4brielvs no worries and thank you for that info! From looking through the discussion, it seems like the reviews have been done and we're just waiting for final sign off? Our editor policy actually allows an editor to step in and affirm that all reviewer concerns have been addressed. Since we're waiting for an update, I'm going to mark this as pending maintainer response and when the new version is published, feel free to comment here and the editor should have the freedom to wrap things up!
